Refactor configuration file functions
